extended_search/python/requirements.txt
lxml
requests
python-dotenv